requirements.txt
# Django
Django==4.2.10
django-filter==24.2
django-dbbackup==4.1.0
dj-database-url==2.1.0
django-lifecycle==1.1.2
django-select2==8.1.2
djangorestframework==3.14.0
django-simple-history==3.4.0
django-smart-selects==1.6.0
django-widget-tweaks==1.4.12
django-import-export==4.0.9
django-autocomplete-light==3.11.0

# Outras Libs
uvicorn==0.34.0
websockets==14.2
fontawesomefree==6.1.2
Faker==22.0.0
daphne==4.1.0
channels==4.0.0
channels-redis==4.2.0
Pillow==9.1.0
psycopg2==2.9.3
python-decouple==3.8
redis==5.0.1
reportlab==4.0.4
pytz==2022.1
sqlparse==0.4.2
zope.interface==6.1
tqdm==4.66.4
xlrd2==1.3.4
weasyprint==62.3
PyPDF2==3.0.1
openpyxl==3.1.5
python-dotenv==1.0.1
sentry-sdk==2.22.0
celery==5.4.0